+++ NEWS.md
@@ -4,6 +4,9 @@
 ### Major:
 * Added `calc_roh` to calculate Runs of Homozygosity and $F_{ROH}$. Produces results very similar to those from "PLINK".
 
+### Minor:
+* Ad
+
 ## Bug fixes:
 * Fixed a bug where on the very newest versions of R setting `scipen` in `options` like `options(scipen = osp)` based on a previous call to `osp <- options(scipen)` would cause an error.
 * Fixed a bug where doing LD calculations in serial without `CLD = only` would skip checking if `bigmemmory` was installed, potentially causing an error.