+++ NEWS.md
@@ -1,15 +1,17 @@
-# snpR 1.3.0
+# snpR 1.2.13
 
 ## Features:
 ### Major:
 * Added `calc_roh` to calculate Runs of Homozygosity and $F_{ROH}$. Produces results very similar to those from "PLINK".
 * Added `plot_roh` to plot the results of `calc_roh`.
+* The newest versions of `data.table` on R 4.5+ on Windows are *very* slow when dcasting with character or factor formula components. The internal tabulation which occurs whenever a new object is made or a sample facet is computed now uses the `pivot` command from `collapse`, which is consistently faster even then `data.table::dcast` using only numeric formula variables. At some point all internal `data.table::dcast` functions will be re-fitted to use `collapse::pivot` for speed.
 
 ## Bug fixes:
 * Fixed a bug where on the very newest versions of R setting `scipen` in `options` like `options(scipen = osp)` based on a previous call to `osp <- options(scipen)` would cause an error.
 * Fixed a bug where doing LD calculations in serial without `CLD = only` would skip checking if `bigmemmory` was installed, potentially causing an error.
 * Attempted to further fix the `data.table` recursion bug from `snpR` 1.2.12, the fix for which apparently didn't work for everyone.
 * Bug fix in `calc_ne` due to a flipped grep call.
+* Fixed a bug in `calc_private` where private alleles could sometimes be attributed to the wrong population when run *without* rarefaction.
 
 # snpR 1.2.12